Renamed householdDTO to response
