+++ backend/src/main/java/stud/ntnu/krisefikser/household/dto/HouseholdResponse.java
